pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mindscout*"]

[project]
name = "mindscout"
version = "0.6.0"
description = "An AI assistant that helps you stay on top of advances in AI"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Kate"}
]
dependencies = [
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "anthropic>=0.40.0",
    "numpy>=1.24.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "slowapi>=0.1.9",
    "apscheduler>=3.10.0",
    # Phoenix observability & evaluation
    "arize-phoenix>=8.0.0",
    "arize-phoenix-evals>=2.0.0",
    "openinference-instrumentation-anthropic>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "coverage[toml]>=7.3.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
mindscout = "mindscout.cli:main"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes (unused imports/variables)
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # function call in default argument (required for FastAPI Depends)
    "B904",   # raise from in except (not needed for HTTPException re-raises)
]

[tool.ruff.lint.isort]
known-first-party = ["mindscout", "backend"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401"]  # unused imports ok in tests (fixtures)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as async",
]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=mindscout",
    "--cov=backend",
    "--cov=mcp-server",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["mindscout", "backend", "mcp-server"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
